Fix Clippy (#1034)
Clean up clippy warnings from `cargo clippy --all-targets`. All changes are in
test code; no production behavior changes.

- vp9 tests: allow erasing_op/identity_op on the unmarshal fixture, use ranges
  instead of hand-rolled counters, drop a redundant `& 0xFF`.
- config: build the inverted MTU range with RangeInclusive::new.
- value_history: rename allow to the new unchecked_time_subtraction lint name.
- handshake-direct: group event-loop I/O into an RtcLoopIo struct, pass
  IceCreds directly, use `if let` in write_pcap, collapse the ICE Checking arm
  into a match guard.
- Assorted repeat_n, next_back, slice::from_ref, needless borrow/cast/clone
  fixes elsewhere.

+++ tests/ice-candidates.rs
@@ -518,7 +518,7 @@ fn ice_stun_max_retransmits() -> Result<(), RtcError> {
     // With max_retransmits=3, we should see limited retransmissions
     // The exact count depends on implementation details, but should be bounded
     // Key verification: count should be <= max_retransmits + 1 (initial + retransmits)
-    let expected_max = (max_retransmits + 1) as usize + 2; // +2 for timing tolerance
+    let expected_max = max_retransmits + 1 + 2; // +2 for timing tolerance
 
     assert!(
         transmit_count <= expected_max,
